[CODE WILL NOT COMPILE] Add 3D model import
Now, a 3D model can be imported in the game engine.

I didn't want to push the code right away (some stuff can still be
improved, and the current state of the code makes it impossible to
be compiled), but I'm too afraid that my computer dies, so I prefer
to be safe, just in case. :)