[WIP] Add fiber-based scheduler
The uindex type has been renamed to usize for improved clarity, and a
ssize type has been added to serve as its signed counterpart.

Additionally, the precompiled header has been removed, as it was not
particularly useful. It has been replaced with an "essentials" header
that contains the minimal code required to work with the engine.